feat: use node weights to track nodes
The implemented Ungraph reshuffles node ids each time a node is removed.
Hence, a method to track the correct nodes is required.
As we do not use node weights, we can now use these as effective node identifiers.
Updated the function `remove_node` and tests to ensure correct trackable behavior.