feat: remove `add_node`
`add_node` is removed as tracking the original size of the circuit is required if we use node weight to track nodes.
The assumption is that we have all available quantum resources from the beginning.
Can be updated in the future for future proofing, but should be good enough for now.