Rename PauliString to PauliVec
The PauliString struct is not really a Pauli string, as it does not
track phase. It is also abused as a column in the CliffordTableau. So a
more generic name is better. As this is a collection of Pauli operators,
PauliVec seemed reasonable.
Of course, this name might be strange to anyone coming across it for the
first time. But I'd argue this is good: If they just read PauliString
and assumed it's a normal Pauli string that would create confusion
later. Instead, we are upfront with that this is not a real Pauli string
and urge the unexperienced to read the documentation to learn about the
purpose of this struct.