Fix possible infinite loop in path_from_shortest_path_tree
The code
while u != v {
   if let Some(x) = ... {

   }
}
will run forever if a None is encountered at a point. Breaking in that
case doesn't make sense because we are returning a partial path without
warning. Since we already checked that self.prev[u][v] is not None, we
expect there exists a path from u to v and, hence, self.prev should also
contain all nodes along this path.