Remove pyenv setup from Makefile
There was nothing project-specific about this setup, just setting up a
normal virtual environment. As there are many tools to do this, I
consider it unnecessary to describe this one tool. It is as if we were
explaining how to install Python. The setup step (or e.g. the .python-version files)
were also not used by the CI.