Bump pyo3 from 0.23.3 to 0.24.2
The old pyo3 version had a minor security issue that Dependabot warned
about. Here's the description:
PyString::from_object took &str arguments and forwarded them directly to
the Python C API without checking for terminating nul bytes. This could
lead the Python interpreter to read beyond the end of the &str data and
potentially leak contents of the out-of-bound read (by raising a Python
exception containing a copy of the data including the overflow).
In PyO3 0.24.1 this function will now allocate a CString to guarantee a
terminating nul byte. PyO3 0.25 will likely offer an alternative API
which takes &CStr arguments.